requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
aiofiles==23.2.1
requests==2.31.0
pandas==2.1.4
beautifulsoup4==4.12.2
selenium==4.15.2
undetected-chromedriver==3.5.4
python-multipart==0.0.6
python-dotenv==1.0.0
